Resolved #137
