+++ src/main/scala/edu/arizona/sista/assembly/RunAssembly.scala
@@ -1,18 +1,10 @@
 package edu.arizona.sista.assembly
 
-import java.io.File
 import com.typesafe.config.ConfigFactory
 import edu.arizona.sista.odin.Mention
-import edu.arizona.sista.reach.nxml.NxmlReader
-import edu.arizona.sista.reach.utils.CSVParser
-import edu.arizona.sista.reach.{context, ReachSystem}
-import edu.arizona.sista.reach.context.ContextEngineFactory.Engine
+import edu.arizona.sista.reach.PaperReader
 import edu.arizona.sista.reach.PaperReader.Dataset
 import edu.arizona.sista.utils.Serializer
-import org.apache.commons.io.FilenameUtils
-import scala.collection.JavaConverters._
-import scala.collection.parallel.ForkJoinTaskSupport
-import scala.collection.parallel.mutable.ParArray
 
 /**
  * Utilities for running assembly sieves on a Dataset and writing their output.
@@ -108,84 +100,10 @@ object RunAssembly extends App {
   val config = ConfigFactory.load()
   val outFolder = config.getString("assembly.outFolder")
 
-  // for context engine
-  val contextEngineType = Engine.withName(config.getString("contextEngine.type"))
-  val contextConfig = config.getConfig("contextEngine.params").root
-  val contextEngineParams: Map[String, String] =
-    context.createContextEngineParams(contextConfig)
-
-  // the number of threads to use for parallelization
-  val threadLimit = config.getInt("threadLimit")
   val papersDir = config.getString("assembly.papers")
-  val ignoreSections = config.getStringList("nxml2fries.ignoreSections").asScala
-
-  // systems for reading papers
-  val nxmlReader = new NxmlReader(ignoreSections)
-  val csvReader = new CSVParser()
-
-  // initialize ReachSystem with appropriate context engine
-  val rs = new ReachSystem(contextEngineType = contextEngineType, contextParams = contextEngineParams)
-
-  def readPapers(path: String): Dataset = readPapers(new File(path))
-
-  /**
-   * Produces Dataset from a directory of nxml and csv papers.
-   * @param dir a directory of nxml and csv papers
-   * @return a Dataset (PaperID -> Mentions)
-   */
-  def readPapers(dir: File): Dataset = {
-    require(dir.isDirectory, s"'${dir.getCanonicalPath}' is not a directory")
-    // read papers in parallel
-    val files = dir.listFiles.par
-    // limit parallelization
-    files.tasksupport =
-      new ForkJoinTaskSupport(new scala.concurrent.forkjoin.ForkJoinPool(threadLimit))
-    // build dataset
-    val data: ParArray[(String, Vector[Mention])] = for {
-      file <- dir.listFiles.par // read papers in parallel
-      // allow either nxml or csv files
-      if file.getName.endsWith(".nxml") || file.getName.endsWith(".csv")
-    } yield file match {
-        case nxml if nxml.getName.endsWith(".nxml") =>
-          readNXMLPaper(nxml)
-        case csv if csv.getName.endsWith(".csv") =>
-          readCSVPaper(csv)
-    }
-    data.seq.toMap
-  }
-
-  /**
-   * Produces Mentions from .csv papers using [[CSVParser]] and [[ReachSystem]]
-   * @param file a File with the .csv extension
-   * @return (PaperID, Mentions)
-   */
-  def readCSVPaper(file: File): (String, Vector[Mention]) = {
-    require(file.getName.endsWith(".csv"), s"Given ${file.getAbsolutePath}, but readCSVPaper only handles .csv files!")
-    val paperID: String = file.getName.replace(".csv", "")
-    println(s"reading paper $paperID ...")
-    val mentions: Seq[Mention] =
-      rs.extractFrom(csvReader.toFriesEntries(file))
-    (paperID, mentions.toVector)
-  }
-
-  /**
-   * Produces Mentions from .nxml papers using [[NxmlReader]] and [[ReachSystem]]
-   * @param file a File with the .csv extension
-   * @return (PaperID, Mentions)
-   */
-  def readNXMLPaper(file: File): (String, Vector[Mention]) = {
-    require(file.getName.endsWith(".nxml"), s"Given ${file.getAbsolutePath}, but readNXMLPaper only handles .nxml files!")
-    val paperID = FilenameUtils.removeExtension(file.getName)
-    println(s"reading paper $paperID ...")
-    val mentions = for {
-      entry <- nxmlReader.readNxml(file)
-      mention <- rs.extractFrom(entry)
-    } yield mention
-    paperID -> mentions.toVector
-  }
 
   // generate Dataset from papers
-  val dataset = readPapers(papersDir)
+  val dataset = PaperReader.readPapers(papersDir)
   // write assembly output files to directory
   writeOutputFromDataset(dataset, outFolder)
 }
